Use Mac OS X 10.5 user space libraries on Mac OS X 10.6 and 10.7
This allows building PowerPC file systems on Mac OS X 10.6 and 10.7 if an
appropriate version of Xcode is installed (e.g. Xcode 3.2.6).

The user space libraries are part of the OSXFUSE core package. There are three
distinct core packages for Mac OS X 10.5, 10.6 and 10.7. At the moment they all
share the same user space libraries which results in a unnecessary large
installer package. This issue will be addressed in OSXFUSE 2.4 and requires
some more or less extensive changes to the build script and the installer
package structure.