Adds mechanism for using a captured .DS_Store file for window settings.
The present setup uses Finder to set up the window parameters for
the install image.  This is undesirable for multiple reasons, and
need not be repeated on every build when the window parameters
haven't changed.

With this change, the .DS_Store file (containing the window
settings) is captured and used for subsequent builds.  The intent
is that a "vetted" copy can be checked into the repository, with
no further need to use Finder in the build until some change is
made to the parameters (at which point the saved version can be
deleted to cause it to be recomputed).

This commit does not include the .DS_Store copy (xDS_Store),
which belongs in a separate commit.  It does include the
binary attribute for xDS_Store.