Sign release instead of individual files
Signing files one at a time signs the contents of the files. However,
it does not protect the file names from being tampered with or prevent
adding or removing files from the release.

Signing a file containing the hash values and names of all release
files effectively signs the release as a whole.